CRM-15 Implement RegistrationForm component, replace SignUpForm, and update authentication flow; enhance Header with navigation links and adjust layout styles.
