+++ src/app/types/auth.interface.ts
@@ -1,4 +1,4 @@
-export interface ISignUpRequest {
+export interface IAuthRequest {
     email: string;
     password: string;
     confirmPassword?: string; // Optional for sign-up forms
@@ -7,4 +7,8 @@ export interface ISignUpRequest {
 export interface ISignUpResponse {
     id: string;
     email: string;
+}
+
+export interface ISignInResponse {
+    accessToken: string;
 }
