+++ src/redux/baseParams.ts
@@ -13,7 +13,6 @@ export const baseParams = fetchBaseQuery({
     
     const token = localStorage.getItem('token');
 
-    // If we have a token set in state, let's assume that we should be passing it.
     if (token) {
       headers.set('authorization', `Bearer ${token}`)
     }
