+++ src/redux/services/auth.ts
@@ -1,15 +1,22 @@
 import { createApi } from "@reduxjs/toolkit/query/react";
 import { authParams } from "../baseParams";
-import { ISignUpRequest, ISignUpResponse } from "@/app/types/auth.interface";
+import { IAuthRequest, ISignInResponse, ISignUpResponse } from "@/app/types/auth.interface";
 
 export const authApi = createApi({
     reducerPath: 'authApi',
     baseQuery: authParams,
     tagTypes: ['User'],
     endpoints: (builder) => ({
-        signUp: builder.mutation<ISignUpResponse, ISignUpRequest>({
+        signUp: builder.mutation<ISignUpResponse, IAuthRequest>({
             query: (userData) => ({
                 url: 'auth/sign-up',
+                method: 'POST',
+                body: userData,
+            }),
+        }),
+        signIn: builder.mutation<ISignInResponse, IAuthRequest>({
+            query: (userData) => ({
+                url: 'auth/sign-in',
                 method: 'POST',
                 body: userData,
             }),
@@ -17,4 +24,4 @@ export const authApi = createApi({
     })
 });
 
-export const { useSignUpMutation } = authApi;
+export const { useSignUpMutation, useSignInMutation } = authApi;
