+++ src/styles/defaultStyles.ts
@@ -1,0 +1,8 @@
+export const defaultStyles = {
+  colors: {
+    gray500: '#6b7280',
+    gray700: '#374151',
+    gray900: '#111827',
+    white: '#ffffff',
+  }
+}
